refactor: change direct links to the constants
