+++ src/shared/helpers/get-schedule-and-documentation-links.tsx
